Refactor redundant autonomous code and increase efficiency
Move the value conversion to the command class, as the original value in rotations or inches was never needed outside of the class. Add addCommands method to the AutoDrive class. Add internal variable to check if commands are running to eliminate the need for placeholder commands. Simplify checking for completion of commands.