Optimize moving average calculation and comment code
The limelight distance calculation was taking the average of an average which was not good. I have removed this.  For the moving average, add a primitive array instead of an arraylist for performance optimization, and also add true averages for the amount of places that are filled in the array.