Remove Nonstants and put respective variables in their proper classes
Nonstants is unneccessary and this is a semi-permanent fix, to move the variables to the classes that the variable most belongs to. This could be later changed with network tables values, but the idea of nonstants as a whole goes against the whole idea of object oriented programming. The variables should go in the classes that abstract that functionality of the robot, even if it is the top level of Robot.java that the variable is eventually put in.